Refactoring: Reusable link and table components
